% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_check_functions.R
\name{add_new_check}
\alias{add_new_check}
\title{This function will add data quality queries and thair status in to a one dataframe}
\usage{
add_new_check(qual_df, Total, check, status, fail_count, fail_percentage)
}
\arguments{
\item{qual_df}{the data frame where to add data quality query and status}

\item{Total}{the denominator of the check}

\item{check}{the check/query}

\item{status}{Pass/Fail status}

\item{fail_count}{fail count, if check pass this is zero}

\item{fail_percentage}{fail percentage, if check pass this is zero}
}
\value{

}
\description{
This function will add data quality queries and thair status in to a one dataframe
}
