% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_check_functions.R
\name{check_date_within}
\alias{check_date_within}
\title{check whether a date is before or after the given date from a dataframe}
\usage{
check_date_within(check_df, check_date, compare_df, compare_date, check_arg)
}
\arguments{
\item{check_df}{an dataframe}

\item{compare_df}{optional. if the comapring date comes from a dataframe}

\item{compare_date}{comparing date or column in a dataframe}

\item{check_arg}{one of the ==, >=,>,<=,<}

\item{check_col}{the column interested within dataset for checking}
}
\value{

}
\description{
check whether a date is before or after the given date from a dataframe
}
