% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_check_functions.R
\name{check_id_availability}
\alias{check_id_availability}
\title{check one column values appear in another column at least once
in a given data frame column}
\usage{
check_id_availability(check_df, check_col, compare_df, compare_col)
}
\arguments{
\item{check_df}{an dataframe we performe check}

\item{check_col}{the column interested within dataset_1}

\item{compare_col}{}
}
\description{
check one column values appear in another column at least once
in a given data frame column
}
