% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_check_functions.R
\name{check_measure_bounds}
\alias{check_measure_bounds}
\title{This function will check whether the given measurement is within the acceptable rate}
\usage{
check_measure_bounds(measurement, bound_df, check_df)
}
\arguments{
\item{measurement}{the measurement variable that need to check}

\item{bound_df}{the data table which have all the measurement bounds}

\item{check_df}{the dataframe}
}
\value{

}
\description{
This function will check whether the given measurement is within the acceptable rate
}
