% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_check_functions.R
\name{find_duplicates}
\alias{find_duplicates}
\title{check duplicates of a given data frame column}
\usage{
find_duplicates(dataset, column)
}
\arguments{
\item{dataset}{an dataframe}

\item{column}{the column interested within ''}
}
\value{

}
\description{
check duplicates of a given data frame column
}
