% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tables.R
\name{get_concept_names}
\alias{get_concept_names}
\title{add concept names for a given table}
\usage{
get_concept_names(df, connection, schema_name)
}
\arguments{
\item{df}{the data frame we need to rename concept ids to names}

\item{connection}{database connection}

\item{schema_name}{character vector of target schema}
}
\description{
add concept names for a given table
}
