% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tables.R
\name{get_period_patients}
\alias{get_period_patients}
\title{Get patients for a selected time period and return as a list
Based on this list all the tables will be filtered
to get the period patients, the observation period table is used}
\usage{
get_period_patients(ctn, schema, start_date, end_date)
}
\arguments{
\item{start_date}{begining of the period}

\item{end_date}{end of the period}

\item{df}{the dataframe which patients selects from}
}
\value{

}
\description{
Get patients for a selected time period and return as a list
Based on this list all the tables will be filtered
to get the period patients, the observation period table is used
}
