% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_check_functions.R
\name{get_source_concept_id}
\alias{get_source_concept_id}
\title{This function will give the concept_id of a given source field in the mapping sheet}
\usage{
get_source_concept_id(measure_field, mapping_sheet)
}
\arguments{
\item{measure_field}{the source or measure field that interested on}

\item{mapping_sheet}{the mapping sheet which have the source field and concept id mappings}
}
\value{

}
\description{
This function will give the concept_id of a given source field in the mapping sheet
}
