% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_visit_detail_profile}
\alias{plot_visit_detail_profile}
\title{Plot Visit detail Profile}
\usage{
plot_visit_detail_profile(x, start_date, end_date, custom_colors)
}
\arguments{
\item{start_date}{starting date for x axis}

\item{end_date}{ending dare for x axis}
}
\value{

}
\description{
Plot Visit detail Profile
}
