% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_large_kable}
\alias{print_large_kable}
\title{Print large kable. Standardises the formatting of full width kables.
Also double checks that we are not printing empty tables.}
\usage{
print_large_kable(table, caption = "", max_rows = 100, print_empty_vars = TRUE)
}
\arguments{
\item{table}{The table to print.}

\item{caption}{Caption for the table}

\item{max_rows}{The maximumn number of rows to print. Defaults to 10.}

\item{print_empty_vars}{Logical indicating if variables with all NAs should be printed}
}
\value{
A neatly formatted full width kable.
}
\description{
Print large kable. Standardises the formatting of full width kables.
Also double checks that we are not printing empty tables.
}
