% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_quality_report}
\alias{render_quality_report}
\title{Render Quality Report}
\usage{
render_quality_report(params = "ask")
}
\arguments{
\item{params}{a string "ask", which will launch an interactive window to enter
database connection details, or a list that contains these details:
\itemize{
\item driver: string, can be any of: "PostgreSQL", "ODBC", or "SQLite"
\item host: string, host name
\item port: integer, port number
\item dbname: string, database name
\item schema: string, database target schema
\item user: string, database username
\item password: string, database user password
\item local_hospital: string, local hospital
}}

\item{output}{an output file path}
}
\description{
Render Quality Report
}
