% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tables.R
\name{get_period_patients}
\alias{get_period_patients}
\title{Get patients for a selected time period and return as a list
Beased on this list all the tables will be filtered}
\usage{
get_period_patients(df, start_date, end_date)
}
\arguments{
\item{df}{the dataframe which patients selects from}

\item{start_date}{begining of the period}

\item{end_date}{end of the period}
}
\value{

}
\description{
Get patients for a selected time period and return as a list
Beased on this list all the tables will be filtered
}
